requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
cryptography==42.0.8
PyJWT==2.8.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1